refacto tileboard
