+++ src/ecs_elements/resources/mod.rs
@@ -3,11 +3,9 @@ pub mod mapgen;
 pub mod menus;
 pub mod assets;
 pub mod sounds;
-pub mod states;
 
 pub use save::ShouldSave;
 pub use mapgen::MapGenHistory;
 pub use menus::MainMenuSelection;
 pub use assets::AsciiSheet;
 pub use sounds::MusicController;
-pub use states::{GameState, AppState};
