+++ src/game/mod.rs
@@ -1,16 +1,24 @@
 // Game Plugin + Component & enum go there + new game setup.
 use bevy::prelude::*;
 
+use self::tileboard::TileBoardPlugin;
 use self::player::PlayerPlugin;
 use self::npc::NpcPlugin;
+
+pub mod player;
+pub mod npc;
+pub mod spawners;
+mod tileboard;
+
+pub use tileboard::components::{Tile, GridPosition};
+
 
 use crate::{
     globals::SHOW_MAPGEN_VISUALIZER,
     map_builders::{
         map::Map, 
-        TileMapPlugin    
     },
-    game::spawners::{spawn_npc, spawn_player},
+    game::{spawners::{spawn_npc, spawn_player}, },
     map_builders::{
         random_builder,
     },    
@@ -20,14 +28,12 @@ use crate::{
     },    
     ecs_elements::{
         components::{Monster},
-        resources::{ShouldSave, MapGenHistory, GameState}, GridPosition,
+        resources::{ShouldSave, MapGenHistory},
     }, 
-    render::GraphicsPlugin
+    render::GraphicsPlugin, states::GameState
 };
 
-pub mod player;
-pub mod npc;
-pub mod spawners;
+
 
 
 
@@ -41,10 +47,10 @@ impl Plugin for GamePlugin {
             .insert_resource(ShouldSave{to_save: false})
             .add_plugin(PlayerPlugin)
             .add_plugin(VictoryPlugin)
-            .add_plugin(TileMapPlugin)
             .add_plugin(NpcPlugin)
             .add_plugin(GameOverPlugin)
             .add_plugin(GraphicsPlugin)
+            .add_plugin(TileBoardPlugin)
 
             .add_systems(OnEnter(GameState::NewGame),init_new_game)
             ;
